+++ frontend/src/pages/ApplicantRegister/ApplicantRegister.js
@@ -3,11 +3,13 @@ import { useHistory, useLocation } from "react-router-dom";
 import {
   BirthField,
   Button,
+  Form,
   GenderField,
   SummaryCheckField,
 } from "../../components/form";
 import RecruitCard from "../../components/RecruitCard/RecruitCard";
 import { POLICY_SUMMARY } from "../../constants/policySummary";
+import useForm from "../../hooks/useForm";
 import useRecruitmentContext from "../../hooks/useRecruitmentContext";
 import useTokenContext from "../../hooks/useTokenContext";
 import FormProvider from "../../provider/FormProvider/FormProvider";
@@ -68,6 +70,20 @@ const ApplicantRegister = () => {
     }
   };
 
+  const { handleSubmit, ...methods } = useForm({
+    validators: {
+      email: validateEmail,
+      name: validateName,
+      password: validatePassword,
+      rePassword: validateRePassword,
+      phoneNumber: validatePhoneNumber,
+      year: validateYear,
+      month: validateMonth,
+      day: validateDay,
+    },
+    submit,
+  });
+
   return (
     <div className={styles["applicant-register"]}>
       {currentRecruitment && (
@@ -77,86 +93,76 @@ const ApplicantRegister = () => {
           endDateTime={currentRecruitment.endDateTime}
         />
       )}
-      <FormProvider
-        submit={submit}
-        validators={{
-          email: validateEmail,
-          name: validateName,
-          password: validatePassword,
-          rePassword: validateRePassword,
-          phoneNumber: validatePhoneNumber,
-          year: validateYear,
-          month: validateMonth,
-          day: validateDay,
-        }}
-      >
-        <h2>지원자 정보</h2>
-        <div>
-          <SummaryCheckField
-            name="policy"
-            label="개인정보 수집 및 이용 동의"
-            required
-          >
-            <p className={styles.summary}>{POLICY_SUMMARY}</p>
-          </SummaryCheckField>
-        </div>
-        <div>
-          <InputField
-            name="name"
-            type="text"
-            label="이름"
-            placeholder="이름을 입력해 주세요."
-            required
-          />
-        </div>
-        <div>
-          <InputField
-            name="phoneNumber"
-            type="text"
-            label="전화번호"
-            placeholder="연락 가능한 전화번호를 입력해 주세요."
-            required
-          />
-        </div>
-        <div>
-          <InputField
-            name="email"
-            type="email"
-            label="이메일"
-            placeholder="이메일 주소를 입력해 주세요."
-            required
-          />
-        </div>
-        <div>
-          <InputField
-            name="password"
-            type="password"
-            label="비밀번호"
-            placeholder="비밀번호를 입력해 주세요."
-            required
-          />
-        </div>
-        <div>
-          <InputField
-            name="rePassword"
-            type="password"
-            label="비밀번호 확인"
-            placeholder="비밀번호를 다시 한번 입력해 주세요."
-            required
-          />
-        </div>
-        <div>
-          <BirthField />
-        </div>
-        <div>
-          <GenderField />
-        </div>
-        <div className={styles["button-wrapper"]}>
-          <Button cancel type="button">
-            취소
-          </Button>
-          <SubmitButton>다음</SubmitButton>
-        </div>
+      <FormProvider {...methods}>
+        <Form onSubmit={handleSubmit}>
+          <h2>지원자 정보</h2>
+          <div>
+            <SummaryCheckField
+              name="policy"
+              label="개인정보 수집 및 이용 동의"
+              required
+            >
+              <p className={styles.summary}>{POLICY_SUMMARY}</p>
+            </SummaryCheckField>
+          </div>
+          <div>
+            <InputField
+              name="name"
+              type="text"
+              label="이름"
+              placeholder="이름을 입력해 주세요."
+              required
+            />
+          </div>
+          <div>
+            <InputField
+              name="phoneNumber"
+              type="text"
+              label="전화번호"
+              placeholder="연락 가능한 전화번호를 입력해 주세요."
+              required
+            />
+          </div>
+          <div>
+            <InputField
+              name="email"
+              type="email"
+              label="이메일"
+              placeholder="이메일 주소를 입력해 주세요."
+              required
+            />
+          </div>
+          <div>
+            <InputField
+              name="password"
+              type="password"
+              label="비밀번호"
+              placeholder="비밀번호를 입력해 주세요."
+              required
+            />
+          </div>
+          <div>
+            <InputField
+              name="rePassword"
+              type="password"
+              label="비밀번호 확인"
+              placeholder="비밀번호를 다시 한번 입력해 주세요."
+              required
+            />
+          </div>
+          <div>
+            <BirthField />
+          </div>
+          <div>
+            <GenderField />
+          </div>
+          <div className={styles["button-wrapper"]}>
+            <Button cancel type="button">
+              취소
+            </Button>
+            <SubmitButton>다음</SubmitButton>
+          </div>
+        </Form>
       </FormProvider>
     </div>
   );
