+++ frontend/src/pages/PasswordEdit/PasswordEdit.js
@@ -9,7 +9,9 @@ import {
   validatePassword,
   validateRePassword,
 } from "../../utils/validation/password";
+import useForm from "../../hooks/useForm";
 import styles from "./PasswordEdit.module.css";
+import { Form } from "../../components/form";
 
 const PasswordEdit = () => {
   const { token, resetToken } = useTokenContext();
@@ -32,44 +34,48 @@ const PasswordEdit = () => {
     }
   };
 
+  const { handleSubmit, ...methods } = useForm({
+    validators: {
+      oldPassword: validatePassword,
+      password: validatePassword,
+      rePassword: validateRePassword,
+    },
+    submit,
+  });
+
   return (
     <div className={styles["password-edit"]}>
-      <FormProvider
-        submit={submit}
-        validators={{
-          oldPassword: validatePassword,
-          password: validatePassword,
-          rePassword: validateRePassword,
-        }}
-      >
-        <h2>비밀번호 변경</h2>
-        <InputField
-          name="oldPassword"
-          type="password"
-          label="기존 비밀번호"
-          placeholder="기존 비밀번호를 입력해 주세요"
-          required
-        />
-        <InputField
-          name="password"
-          type="password"
-          label="새 비밀번호"
-          placeholder="비밀번호를 입력해 주세요"
-          required
-        />
-        <InputField
-          name="rePassword"
-          type="password"
-          label="비밀번호 확인"
-          placeholder="비밀번호를 다시 한 번 입력해 주세요"
-          required
-        />
-        <div className={styles.buttons}>
-          <Button cancel onClick={() => history.goBack()}>
-            이전
-          </Button>
-          <SubmitButton>확인</SubmitButton>
-        </div>
+      <FormProvider {...methods}>
+        <Form onSubmit={handleSubmit}>
+          <h2>비밀번호 변경</h2>
+          <InputField
+            name="oldPassword"
+            type="password"
+            label="기존 비밀번호"
+            placeholder="기존 비밀번호를 입력해 주세요"
+            required
+          />
+          <InputField
+            name="password"
+            type="password"
+            label="새 비밀번호"
+            placeholder="비밀번호를 입력해 주세요"
+            required
+          />
+          <InputField
+            name="rePassword"
+            type="password"
+            label="비밀번호 확인"
+            placeholder="비밀번호를 다시 한 번 입력해 주세요"
+            required
+          />
+          <div className={styles.buttons}>
+            <Button cancel onClick={() => history.goBack()}>
+              이전
+            </Button>
+            <SubmitButton>확인</SubmitButton>
+          </div>
+        </Form>
       </FormProvider>
     </div>
   );
